Make interp into a Coq function of the environment
Since the environment is basically discrete anyway, a Coq function from it is
equivalent to a non-expansive function from its OFE.
However, it's not obvious to the tactics that listC vlC is discrete if vlC is
discrete, so non-expansiveness is not trivial to prove.

This also speeds up the build!