Extract definition typing from member typing, plus automation
We'll soon want to add semantic definition typing and relate it to semantic
value typing. So, split out what I'm gonna need for definitions.

Proofs of affected semantic lemmas needed to be almost rewritten.

Some trivial Coq proof steps can be automated, and automation reduces the work
to deal with such trivial changes, so import (some of) my homegrown tactics for
that.