Finally store *open* semantic types (list vl * vl -> iProp Σ)
Most of the work about the translation is still left to do.

I made up a function to generate an identity σ, and asserted it's the
appropriate identity; that lemma is not a trivial induction and I ran out of
steam, tho at least I figured why it's true. But it might well be unnecessary,
since it never shows up on paper.