Deindex syntactic typing judgment
Thanks to changes to definition typing, we needn't use level one for typing
definition values.

Do beware that the enabling changes loose precision, but I think in an
acceptable and standard way.
Using on-paper syntax, consider proving definition typing for a member value,
specifically, proving `Γ | this : T ⊢ { a = v } : { a : U }`.

- Without enabling commits, we must prove (something equivalent to) `Γ, this:
  TLater T ⊢ v : TLater U`.

- With enabling commits, we must prove `Γ, this: T ⊢ v : U`: this discards
  the fact that `Γ` is in fact typed in an earlier world.
  However, that's standard for tuples when the product type is contractive, and
  removes an annoying index from the typing rules.