Make `to_subst` transparent
Drop-in change.
Still prevent simplification, but allow unification to see through it.
This is necessary to prepare for #80 and #83, otherwise e.g. calls to e.g.
`iApply ("Hsub" $! (v :: ρ))` must be turned into calls to:

```coq
iSpecialize ("Hsub" $! (v :: ρ)); rewrite to_subst_cons; iApply "Hsub".
```

While at it, prepare to allow simplifying `to_subst` (in next commit).
Improve simplification behavior of `to_subst` by defining it as a
fixpoint (equivalent to the current definition): we don't want to see `foldr`
when it's reduced.