Functorize again SortsSig
This undos parts of 2f41b88f6f299b87426f489bd8a24455b034a8c2.

However, instead of having modules that depend on both ValueSig and SortsSig, we
make them depend on a combined signature.

This way we avoid parameterizing on a potentially high number of functor
arguments, if we add more abstraction in this style.