Hide plength from lemmas (breaks fundamental.v)
Some lemmas are now "loose", but I think that's interesting enough to look at.

Earlier: WIP on changing ipwp to hide plength, cannot work?

The problem is that `Sub_Sel_Path` and especially `Sel_Sub_Path` are asymmetric.
Ignoring extra weakening, changing the level of the premise cannot adjust the
level of *one side* of the conclusion!