Drop binding side condition; blocked
This commit modifies `extraction`, `T ~[ n ] (g, (s, σ))`, to not require
closure of `σ`.

This is blocked, as we need to drop side conditions for term closedness.

Also, this might be a bad idea, as we do still need sometimes closedness of
types, and we do lose some results about it.

However, other simplifications here might be useful.