WIP success (up to skip) not compiling
Apparently, instead of having Sub-Later-Mono as a primitive, one
can make it admissible (as it should be).
However, one needs to add:
```coq
| Path_var_typed x T i :
  Γ u⊢ₚ x : T, i →
  Γ u⊢ₜ x : iterate TLater i T
```
to be able to prove
```coq
  Γ u⊢ₜ T1, i <: T2, j →
  TLater <$> Γ u⊢ₜ T1, S i <: T2, S j
```

And moreover one must add:
```
| undelay_stp Γ Γ' T1 T2 i j :
    ⊢G Γ <:* Γ' →
    Γ' u⊢ₜ T1, i <: T2, j →
    Γ  u⊢ₜ T1, i <: T2, j.
```
to deduce:
```coq
  Γ u⊢ₜ T1, i <: T2, j →
  TLater <$> Γ u⊢ₜ T1, S i <: T2, S j
```

This is only sketched, and I already expect some extra side conditions in rules,
and I'm sure I'll need extra hypotheses for the admissible rules.

Indeed, I'll have to port stamped_mut_types to finish this.
Or I'll just prove admissibility for stamped/storeless typing?