Optimization: Make `dlangG_persistent` a direct proof of `Persistent P`
Now proving `Persistent P` takes negligible time (and produces a slightly
smaller term).

That's the style I used with `SwapPropI`: proving persistence from camera
properties (or anything else!) is a job for the adequacy proof.

This might also make the previous tuning of priorities unnecessary, but it
can't hurt, so leaving that in.