Tweak `ipwp_terminates` for later credits: don't rely on `bupd_plain`
Usually one cannot `iMod` a basic update without one in the goal; this only
works thanks to `elim_modal_bupd_plain_goal`, which relies on `bupd_plain`.

Instead, we can just add a basic update to the goal. This is fine because the
only user is `ipwp_gs_adequacy`, which already accepts this basic update, and
eliminates it explicitly via `bupd_plain_soundness` (which also relies on
`bupd_plain`).

Another motivation is that later credits do not offer `bupd_plain` as is, only
`lc_soundness`, so this step would need adjusting if we used `le_upd` or fancy
updates. Hence, it's worth making this step explicit.